Move pie plot onPlotlyRender call to FacetedPiePlot
